backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1
pymongo==4.10.1
python-dotenv==1.0.1
pytest==8.3.5
httpx==0.28.1
